bug de botón solucionado y componente menu.js para no repetir código con breakfast y lunch
