+++ package.json
@@ -18,7 +18,8 @@
     "react-bootstrap": "^1.0.0-beta.12",
     "react-dom": "^16.9.0",
     "react-router-dom": "^5.1.0",
-    "react-scripts": "3.1.1"
+    "react-scripts": "3.1.1",
+    "react-uuid": "^1.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
