+++ src/components/Home.js
@@ -1,6 +1,5 @@
 import React from "react";
 import "../App.css";
-//import logo from
 import { Link } from "react-router-dom";
 
 const Home = () => (
