+++ src/components/button.css
@@ -6,7 +6,7 @@ item-btn {
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
     margin-left: 5px;
     padding: 3px;
-    width: 200%
+    width: 200%;
   }
 
 .item-btn:hover {
